backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
streamlit==1.31.0
pyautogui==0.9.54
opencv-python==4.8.1
numpy==1.24.3
Pillow==8.3.1
mss==6.1.0 